requirements/base.txt
# Wheel 0.25+ needed to install certain packages on CPython 3.5+
# like Pillow and psycopg2
# See http://bitly.com/wheel-building-fails-CPython-35
# Verified bug on Python 3.5.1
wheel==0.29.0

# Bleeding edge Django
django==1.9.9

# Configuration
django-environ==0.4.0



# Forms
django-braces==1.9.0
django-crispy-forms==1.6.0

# Models
django-model-utils==2.5.2

# Images
Pillow==3.3.1

# For user registration, either via email or social
# Well-built with regular release cycles!
django-allauth==0.27.0


# Python-PostgreSQL Database Adapter
psycopg2==2.6.2

# Unicode slugification
awesome-slugify==1.6.5
django-autoslug==1.9.3

# Time zones support
pytz==2016.6.1

# Redis support
django-redis==4.4.4
redis>=2.10.0

# drf
djangorestframework==3.4.7
markdown==2.6.6
django-filter==0.15.0
drfdocs==0.0.11

django-cors-headers==1.2.0
